moved coursedata into plugin.py
